All is OK

+++ src/character/services/character.service.ts
@@ -4,12 +4,14 @@ import { Repository, UpdateResult } from 'typeorm';
 import { Character } from '../character.entity';
 import { CreatedCharacterDto } from '../dto/character.dto';
 import { UpdatedCharacterDto } from '../dto/updatedCharacter.dto';
+import { CharactersItemService } from 'src/characterItem/services/characterItem.service';
 
 @Injectable()
 export class CharactersService {
     constructor(
         @InjectRepository(Character)
         private charactersRepository: Repository<Character>,
+        private charactersItemService: CharactersItemService,
     ) {}
     
     async GetAll(): Promise<Character[]> {
@@ -27,6 +29,9 @@ export class CharactersService {
 
     async Delete(id: number): Promise<Character[]> {
         let character = await this.charactersRepository.findOne({ where: {id: id} });
+        if (!character) {
+            return null;
+        }
         return this.charactersRepository.remove([character]);
     }
